requirements.txt
pypdf==5.5.0
chromadb==1.0.12
anthropic==0.18.1
streamlit==1.45.1
numpy==2.2.6
python-dotenv==1.1.0
PyPDF2==3.0.1
openai==1.82.1
tiktoken==0.9.0